pyproject.toml
[project]
name = "infra-automation-mcp"
version = "1.0.0"
description = "MCP server for AI-powered infrastructure automation - Okta, Terraform, AWS, and GitHub"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "PyGithub>=2.1.0",
]

[project.scripts]
infra-automation-mcp = "infra_automation_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/infra_automation_mcp"]